smalltools/docTosql/requirements.txt
cryptography
mysql-connector-python
pandas
sqlalchemy
xlrd